requirements.txt
gradio
transformers
torch
google-generativeai